Move process-wide refresh lock state into its own module
refresh_runner.py held two unrelated concerns: dispatching a knowledge
refresh (subprocess spawn, state reconciliation, publish decisions) and
maintaining the process-wide tables that serialize refresh work and count
active refreshes.

Move the second concern to knowledge/refresh_locks.py: the lock table and
its guard, the active-refresh counts and their guard, the entry dataclass,
the borrow/release/prune helpers, the in-loop and cross-process lock
context managers, and the four activity predicates. Every moved body is
unchanged; the two lock context managers lose their leading underscore
because they are now imported across modules.

The lock tables are process-wide singletons, so callers must reach exactly
one copy. All importers now point at the owning module, and Tach limits
visibility to the three modules that legitimately need it.
